% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse}
\alias{fuse}
\title{Combine FUSE soil moisture accounting module and routing module in one function}
\usage{
fuse(DATA, mid, deltim, ParameterSet)
}
\arguments{
\item{DATA}{data.frame containing observations. It consists of 3 columns:
Rainfall (P), Potential Evapo-Transpiration (E) and Streamflow (Q)}

\item{mid}{model id number in Model List 2011(see below for details)}

\item{deltim}{observation time step (days)}

\item{ParameterSet}{list of parameters}
}
\value{
Simulated streamflow discharge
}
\description{
Combine FUSE soil moisture accounting module and routing module in one function
}
\details{
The list of parameters can be generated as follows: ParameterSet <- GeneratePsetsFUSE(1).
}
\examples{
\dontrun{
data(DATA)
set.seed(123)
parameters <- generateParameters(1)
Q <- fuse(DATA, 60, 1, parameters)
}

}

