% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-data.R
\docType{data}
\name{fuse_hydrological_timeseries}
\alias{fuse_hydrological_timeseries}
\title{FUSE dummy dataset}
\format{The format is: 'zoo' series from 2000-01-01 to 2002-12-31
The data consists of a numeric data.frame with 1096 rows and 3 columns named: "P", "E" and "Q". The temporal index consist of daily dates from "2000-01-01" to "2000-01-05".}
\description{
This dataset consists of a data.frame with 3 columns: P (precipitation time series), E(potential evapo-transpiration time series), Q(discharge time series).
}
\details{
Data set: fuse_hydrological_timeseries
}
\examples{
data(fuse_hydrological_timeseries)
str(fuse_hydrological_timeseries)
plot(fuse_hydrological_timeseries)

}
\keyword{datasets}
