% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuseinfo.R
\name{FUSEinfo}
\alias{FUSEinfo}
\title{This function returns information on model structure components and parameters used, given FUSE model.}
\usage{
FUSEinfo(mid)
}
\arguments{
\item{mid}{FUSE model structure ID number (integer from 1 to 1248).}
}
\description{
This function returns information on model structure components and parameters used, given FUSE model.
}
\examples{
# FUSEinfo(mid=60)

}

