% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuserouting.R
\name{fuserouting.ranges}
\alias{fuserouting.ranges}
\title{Function to define the parameter ranges for GAMMA routing module}
\usage{
fuserouting.ranges()
}
\value{
The function returns the list of parameter ranges for the Routing model.
}
\description{
Parameter ranges for GAMMA routing module, as suggested in Clark et al. (2011).
}
\examples{
\dontrun{
fuserouting.ranges()
}

}
