% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuserouting.R
\name{fuserouting.sim}
\alias{fuserouting.sim}
\title{FUSE Routing model}
\usage{
fuserouting.sim(U, mid, deltim, timedelay)
}
\arguments{
\item{U}{This is the effective rainfall/instantaneous runoff, which is the sum of surface runoff, overflow, interflow, and baseflow. This vector can either be of class zoo or numeric.}

\item{mid}{This is the model identification number (see first column of \code{modlist}). This is a numeric value in the range [1, 1248].}

\item{deltim}{This is the input time step (numeric value). By default deltim = 1 for daily steps. All options: deltim = 1 (daily time step), 1/24 (hourly time step), 1/24/4 (15 min time step).}

\item{timedelay}{Time delay in runoff (days). This is a numeric value.}
}
\value{
The function returns an array of simulated "routed" discharges. It can be used after calculating instantaneous discharges with \code{fusesma.sim}.
}
\description{
Implementation of the Routing module of the framework for hydrological modelling FUSE, derived from the Fortran version of FUSE by Martyn Clark (2011).
}
\details{
fuserouting.sim() is a routing module based on a two parameter Gamma distribution. It takes in input the instantaneous discharge and returns the routed discharge. It is compatible with the HYDROMAD framework (see hydromad package: http://hydromad.catchment.org/). For more information on suggested parameter ranges see Clark et al. 2011. Also see \code{GenerateFUSEParameters}.
}
\examples{
\dontrun{
data(fuse_hydrological_timeseries)
set.seed(123)
parameters <- generateParameters(1)
U <- fusesma.sim(fuse_hydrological_timeseries,
                 60, 1, parameters$rferr_add, parameters$rferr_mlt,
                 parameters$frchzne, parameters$fracten,
                 parameters$maxwatr_1, parameters$percfrac,
                 parameters$fprimqb, parameters$qbrate_2a,
                 parameters$qbrate_2b, parameters$qb_prms,
                 parameters$maxwatr_2, parameters$baserte,
                 parameters$rtfrac1, parameters$percrte, parameters$percexp,
                 parameters$sacpmlt, parameters$sacpexp, parameters$iflwrte,
                 parameters$axv_bexp, parameters$sareamax,
                 parameters$loglamb, parameters$tishape, parameters$qb_powr)
Q <- fuserouting.sim(U, 60, 1, 0.5)
}

}
\references{
{
Clark M. P., SlaterA. G., Rupp D. E., Woods R. A., Vrugt J. A., Gupta H. V., Wagener T. and Hay L. E. (2008), Framework for Understanding Structural Errors (FUSE): A modular framework to diagnose differences between hydrological models, Water Resour. Res. 44 p. 91-94.
Clark M. P., McMillan H. K., Collins D. B. G., Kavetski D. and Woods R. A. (2011), Hydrological field data from a modeller's perspective: Part 2: process-based evaluation of model hypotheses. Hydrological Processes, 25: 523-543. doi: 10.1002/hyp.7902.
}
}
