% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GenerateFUSEParameters}
\alias{GenerateFUSEParameters}
\title{This function generates parameter sets for FUSE (Clark et al., 2008).}
\usage{
GenerateFUSEParameters(NumberOfRuns, SamplingType = "LHS", rferr_add = NULL,
  rferr_mlt = NULL, maxwatr_1 = NULL, maxwatr_2 = NULL, fracten = NULL,
  frchzne = NULL, fprimqb = NULL, rtfrac1 = NULL, percrte = NULL,
  percexp = NULL, sacpmlt = NULL, sacpexp = NULL, percfrac = NULL,
  iflwrte = NULL, baserte = NULL, qb_powr = NULL, qb_prms = NULL,
  qbrate_2a = NULL, qbrate_2b = NULL, sareamax = NULL, axv_bexp = NULL,
  loglamb = NULL, tishape = NULL, timedelay = NULL)
}
\arguments{
\item{NumberOfRuns}{number of samples to generate, can be any integer}

\item{SamplingType}{sampling procedure to use, can be "URS" or "LHS"}

\item{rferr_add}{range of the additive rainfall error (mm), default is c(0,0)}

\item{rferr_mlt}{range of the multiplicative rainfall error (-), default is c(1,1)}

\item{maxwatr_1}{range of the depth of the upper soil layer (mm), default is c(25,500)}

\item{maxwatr_2}{range of the depth of the lower soil layer (mm), default is c(50,5000)}

\item{fracten}{range of the fraction total storage in tension storage (-), default is c(0.05,0.95)}

\item{frchzne}{range of the fraction tension storage in recharge zone (-), default is c(0.05,0.95)}

\item{fprimqb}{range of the fraction storage in 1st baseflow reservoir (-), default is c(0.05,0.95)}

\item{rtfrac1}{range of the fraction of roots in the upper layer (-), default is c(0.05,0.95)}

\item{percrte}{range of the percolation rate (mm day-1), default is c(0.01,1000)}

\item{percexp}{range of the percolation exponent (-), default is c(1,20)}

\item{sacpmlt}{range of the SAC model percltn mult for dry soil layer (-), default is c(1,250)}

\item{sacpexp}{range of the SAC model percltn exp for dry soil layer (-), default is c(1,5)}

\item{percfrac}{range of the fraction of percltn to tension storage (-), default is c(0.05,0.95)}

\item{iflwrte}{range of the interflow rate (mm day-1), default is c(0.01,1000)}

\item{baserte}{range of the baseflow rate (mm day-1), default is c(0.001,1000)}

\item{qb_powr}{range of the baseflow exponent (-), default is c(1,10)}

\item{qb_prms}{range of the baseflow depletion rate (day-1), default is c(0.001,0.25)}

\item{qbrate_2a}{range of the baseflow depletion rate 1st reservoir (day-1), default is c(0.001,0.25)}

\item{qbrate_2b}{range of the baseflow depletion rate 2nd reservoir (day-1), default is c(0.001,0.25)}

\item{sareamax}{range of the maximum saturated area (-), default is c(0.05,0.95)}

\item{axv_bexp}{range of the ARNO/VIC b exponent (-), default is c(0.001,3)}

\item{loglamb}{range of the mean value of the topographic index (m), default is c(5,10)}

\item{tishape}{range of the shape param for the topo index Gamma dist (-), default is c(2,5)}

\item{timedelay}{range of the time delay in runoff (days), default is c(0.01,5)}
}
\description{
This function generates parameter sets for FUSE (Clark et al., 2008).
}
\examples{
# For reproducible results, use set.seed() before running this function.
# set.seed(123)
# parameters <- GenerateFUSEParameters(NumberOfRuns=1000)
}

