% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_param_table.R
\name{model_param_table}
\alias{model_param_table}
\title{Prepare a table containing Model ID numbers, model options and parameter in use.}
\usage{
model_param_table(reduced = FALSE)
}
\arguments{
\item{reduced}{This is a logical value. If FALSE the entire model list is taken into account. If FALSE, a reduced version is selected (78 models, according to Clark et al. (2008)).}
}
\value{
table containing Model ID numbers, model options and parameter in use
}
\description{
Prepare a table containing Model ID numbers, model options and parameter in use.
}
\examples{
# model_param_table()

}

