% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readmd}
\alias{readmd}
\title{Read Model Decisions}
\usage{
readmd(mid)
}
\arguments{
\item{mid}{model id number in Model List 2011(see below for details)}
}
\value{
prints on the screen a desciption of the selected model
}
\description{
Read Model Decisions
}

