% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readmd2var}
\alias{readmd2var}
\title{Read Model Decisions}
\usage{
readmd2var(mid, number = FALSE, text = TRUE)
}
\arguments{
\item{mid}{model id number in Model List 2011(see below for details)}

\item{number}{boolean value. If set to TRUE, the output is in num format.}

\item{text}{boolean value. If set to TRUE, the output is in text format.}
}
\value{
prints on the screen a desciption of the selected model
}
\description{
Read Model Decisions
}

