% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RunFUSE}
\alias{RunFUSE}
\title{Combine FUSE soil moisture accounting module and routing module in one function}
\usage{
RunFUSE(DATA, ParameterSet, deltim, mid)
}
\arguments{
\item{DATA}{data.frame containing observations. It consists of 3 columns:
Rainfall (P), Potential Evapo-Transpiration (E) and Streamflow (Q)}

\item{ParameterSet}{list of parameters}

\item{deltim}{observation time step (days)}

\item{mid}{model id number in Model List 2011(see below for details)}
}
\value{
Simulated streamflow discharge
}
\description{
Combine FUSE soil moisture accounting module and routing module in one function
}
\details{
The list of parameters can be generated as follows: ParameterSet <- GeneratePsetsFUSE(1).
}

