% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-package.R, R/fuse.R
\docType{package}
\name{fuse}
\alias{fuse}
\alias{fuse-package}
\title{fuse: Framework for Understanding Structural Errors}
\usage{
fuse(DATA, mid, deltim, ParameterSet)
}
\arguments{
\item{DATA}{data.frame containing observations. It consists of 3 columns:
Rainfall (P), Potential Evapo-Transpiration (E) and Streamflow (Q)}

\item{mid}{model id number in Model List 2011(see below for details)}

\item{deltim}{observation time step (days)}

\item{ParameterSet}{list of parameters}
}
\value{
Simulated streamflow discharge
}
\description{
Implementation of the framework for hydrological modelling FUSE, based on the Fortran version described in Clark et al. (2008). The package consists of two modules: Soil Moisture Accounting module (\code{fuse.sim}) and Gamma routing module (\code{routgamma.sim}). It also contains default parameter ranges (see \code{fuse.ranges} and \code{routgamma.ranges}) and three data objects: DATA, parameters and modlist.

Combine FUSE soil moisture accounting module and routing module in one function
}
\details{
The list of parameters can be generated as follows: ParameterSet <- GeneratePsetsFUSE(1).
}
\examples{
\dontrun{
data(DATA)
set.seed(123)
parameters <- generateParameters(1)
Q <- fuse(DATA, 60, 1, parameters)
}

}

