% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuseInfo.R
\name{fuseInfo}
\alias{fuseInfo}
\title{This function returns information on model structure components and parameters used, given FUSE model.}
\usage{
fuseInfo(mid)
}
\arguments{
\item{mid}{FUSE model structure ID number (integer from 1 to 1248).}
}
\description{
This function returns information on model structure components and parameters used, given FUSE model.
}
\examples{
\dontrun{
  fuseInfo(mid=60)
}

}

