% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuseInfo.R
\name{fuseInfo}
\alias{fuseInfo}
\title{Generate information on model structure components and parameter values}
\usage{
fuseInfo(mid)
}
\arguments{
\item{mid}{FUSE model structure ID number (integer from 1 to 1248).}
}
\description{
fuseInfo is used to generate tables to summarise which model component and parameters are used given the model id. This is important for calculating the posterior distribution of parameter values and model structures.
}
\examples{
\dontrun{
  fuseInfo(mid=60)
}

}

