\name{fuserouting.ranges}
\alias{fuserouting.ranges}
\title{Function to define the parameter ranges for GAMMA routing module}
\usage{
fuserouting.ranges()
}
\description{
Parameter ranges for GAMMA routing module, as suggested in Clark et al. (2008)
}
\details{
Function to define the parameter ranges for GAMMA routing module
}
\author{
Claudia Vitolo
}

