\name{fusesma.ranges}
\alias{fusesma.ranges}
\title{Function to define the parameter ranges for FUSE Soil Moisture Accounting module}
\usage{
fusesma.ranges()
}
\description{
Parameter ranges for FUSE Soil Moisture Accounting module, as suggested in Clark et al. (2008)
}
\details{
Function to define the parameter ranges for FUSE Soil Moisture Accounting module
}
\author{
Claudia Vitolo
}

