% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-data.R
\docType{data}
\name{modlist}
\alias{modlist}
\title{FUSE model structures (numeric)}
\format{A data frame with 1248 rows and 9 variables}
\description{
This object contains the list of FUSE model structures implemented by Clark et al. (2011). This is available as numerical codes (\code{modlist}) or strings (\code{modliststring}). The variables are as follows:
\itemize{
  \item mid = model id number
  \item rferr = type of rainfall error
  \item arch1 = architecture of the upper soil layer
  \item arch2 = architecture of the lower soil layer
  \item qsurf = surface runoff
  \item qperc = vertical drainage
  \item esoil = evapotranspiration
  \item qintf = interflow
  \item q_tdh = routing
}
Each row identifies a model structure and each column identifies a different element of the model structure. For each element of the model structure, several model decisions can be made (see Clark et al. 2011 for details):
\itemize{
  \item Rainfall error (rferr) \itemize{
      \item additive_e (additive rainfall error) = \strong{11}
      \item multiplc_e (multiplicative rainfall error) = \strong{12}
      }
  \item Architecture of the upper soil layer (arch1) \itemize{
      \item onestate_1 (Single state: upper layer defined by a single state variable) = \strong{21}
      \item tension1_1 (Separate tension storage: upper layer broken up into tension and free storage) = \strong{22}
      \item tension2_1 (Cascading buckets: tension storage sub-divided into recharge and excess) = \strong{23}
      }
  \item Architecture of the lower soil layer (arch2) \itemize{
      \item fixedsiz_2 (baseflow reservoir of fixed size) = \strong{31}
      \item tens2pll_2 (tension reservoir plus two parallel tanks) = \strong{32}
      \item unlimfrc_2 (baseflow reservoir of unlimited size (0-HUGE), frac rate) = \strong{33}
      \item unlimpow_2 (baseflow reservoir of unlimited size (0-HUGE), power recession) = \strong{34}
      %% \item topmdexp_2 (topmodel exponential reservoir (no more in the list) = \strong{35}
   }
   \item Runoff (qsurf) \itemize{
       \item arno_x_vic (Unsaturated zone Pareto: ARNO/Xzang/VIC parameterization (upper zone control)) = \strong{41}
       \item prms_varnt (Unsaturated zone linear: PRMS variant (fraction of upper tension storage)) = \strong{42}
       \item tmdl_param (Saturated zone topographic: TOPMODEL parameterization (only valid for TOPMODEL qb)) = \strong{43}
       }
   \item Percolation (qperc) \itemize{
       \item perc_f2sat (Drainage above field capacity: water from (field cap to sat) avail for percolation) = \strong{51}
       \item perc_lower (Gravity drainage: percolation defined by moisture content in lower layer (SAC)) = \strong{52}
       \item perc_w2sat (Saturated zone control: water from wilt point to saturation is available for percolation) = \strong{53}
       }
  \item Evaporation (esoil) \itemize{
      \item rootweight (root weighting) = \strong{61}
      \item sequential (sequential evaporation model) = \strong{62}
   }
   \item Interflows (qintf) \itemize{
       \item intflwnone (interflow denied) = \strong{71}
       \item intflwsome (interflow allowed) = \strong{72}
   }
   \item Routing (q_tdh) \itemize{
       \item no_routing (routing denied) = \strong{81}
       \item rout_gamma (routing allowed using Gamma distribution with shape parameter = 2.5) = \strong{82}
       }
 }
 The function \code{readmd()} returns the model components corresponding to a given id. A similar function (\code{readmd2var()}) reads the components as either character or numeric vector.

On Clark's model list, the parent models (Topmodel, VIC, PRMS and Sacramento) have id = 60, 230, 342 and 426 respectively.

For example, model 5 is identified by the following numeric codes:
data(modlist)
modlist[5,]
}
\details{
Data set: modlist
}

