\name{modlist}
\docType{data}
\alias{modlist}
\title{List of model structures as per Clark et al., 2011}
\description{
'modlist' is the list of model structures implemented by Clark et al. (2011).
}
\usage{data}
\format{data frame}
%\source{}
\references{
Clark M. P., SlaterA. G., Rupp D. E., Woods R. A., Vrugt J. A., Gupta H. V., Wagener T. and Hay L. E. (2008), Framework for 	Understanding Structural Errors (FUSE): A modular framework to diagnose differences between hydrological models, Water Resour. Res. 44 p. 91-94

Clark M. P., McMillan H. K., Collins D. B. G., Kavetski D. and Woods R. A. (2011), Hydrological field data from a modeller's perspective: Part 2: process-based evaluation of model hypotheses. Hydrological Processes, 25: 523-543. doi: 10.1002/hyp.7902 
}
\keyword{datasets}

