% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmd.R
\name{readmd}
\alias{readmd}
\title{Read Model Decisions}
\usage{
readmd(mid)
}
\arguments{
\item{mid}{model id number in Model List 2011(see below for details)}
}
\value{
prints on the screen a description of the selected model
}
\description{
readmd is used for reporting, to generate tables (usually via xtable) to list model structures and describe each component.
}
\examples{
\dontrun{
# To read the model components corresponding to a given id:
readmd(mid = 5)
}

}

